+++ src/test/java/org/jabref/logic/crawler/CrawlerTest.java
@@ -16,6 +16,7 @@ import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.metadata.SaveOrderConfig;
 import org.jabref.model.util.DummyFileUpdateMonitor;
+import org.jabref.preferences.PreferencesService;
 
 import org.eclipse.jgit.api.Git;
 import org.junit.jupiter.api.Test;
@@ -34,6 +35,7 @@ class CrawlerTest {
     @TempDir
     Path tempRepositoryDirectory;
     ImportFormatPreferences importFormatPreferences;
+    PreferencesService preferencesService;
     SavePreferences savePreferences;
     TimestampPreferences timestampPreferences;
     BibEntryTypesManager entryTypesManager;
@@ -48,7 +50,7 @@ class CrawlerTest {
         Crawler testCrawler = new Crawler(getPathToStudyDefinitionFile(),
                 gitHandler,
                 new DummyFileUpdateMonitor(),
-                importFormatPreferences,
+                preferencesService,
                 savePreferences,
                 timestampPreferences,
                 entryTypesManager
@@ -91,9 +93,11 @@ class CrawlerTest {
                 GlobalCitationKeyPattern.fromPattern("[auth][year]"),
                 ',');
 
+        preferencesService = mock(PreferencesService.class);
         importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         savePreferences = mock(SavePreferences.class, Answers.RETURNS_DEEP_STUBS);
         timestampPreferences = mock(TimestampPreferences.class);
+        when(preferencesService.getImportFormatPreferences()).thenReturn(importFormatPreferences);
         when(savePreferences.getSaveOrder()).thenReturn(new SaveOrderConfig());
         when(savePreferences.getEncoding()).thenReturn(null);
         when(savePreferences.takeMetadataSaveOrderInAccount()).thenReturn(true);