+++ src/test/java/org/jabref/logic/importer/fetcher/IEEETest.java
@@ -11,6 +11,7 @@ import org.jabref.logic.importer.SearchBasedFetcher;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.types.StandardEntryType;
+import org.jabref.preferences.PreferencesService;
 import org.jabref.testutils.category.FetcherTest;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -29,8 +30,10 @@ class IEEETest implements SearchBasedFetcherCapabilityTest, PagedSearchFetcherTe
     @BeforeEach
     void setUp() {
         ImportFormatPreferences importFormatPreferences = mock(ImportFormatPreferences.class);
+        PreferencesService preferencesService = mock(PreferencesService.class);
         when(importFormatPreferences.getKeywordSeparator()).thenReturn(',');
-        fetcher = new IEEE(importFormatPreferences);
+        when(preferencesService.getImportFormatPreferences()).thenReturn(importFormatPreferences);
+        fetcher = new IEEE(preferencesService);
         entry = new BibEntry();
     }